requirements.txt
# Heroku deploy
-r requirements-heroku.txt

# Azure deploy
# -r requirements-azure.txt

# Minimum Django and REST framework version
Django==1.10
djangorestframework==3.5.3
django-filter==1.0.1
Markdown==2.6.7

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1

# Swagger Django Rest Framework Documentation
django-rest-swagger

Jinja2==2.8
MarkupSafe==0.23
PyYAML==3.12
argh==0.26.2
dj-database-url==0.4.1
dj-static==0.0.6
django-extensions==1.7.5
ghp-import==0.4.1
mccabe==0.5.2
pathtools==0.1.2
py==1.4.31
python-decouple==3.0
pytz==2016.10
six==1.10.0
static3==0.7.0
watchdog==0.8.3

